Add SQLite database instead of holding all elements in ArrayList
NOTE: The Maven dependency for the sqlite jdbc driver does not link to
the project automatiaclly, therefore, a lib directory with the driver
has been added, and may need to be linked to the project's build path.

- Added SQLite database. This database will be updated each time a file
is loaded, and will have the option to refresh the currently opened
file.
- Added id fields to all elements. Some of the elements maintain their
old id fields.